# docs/svc/openapi.yaml
components:
    schemas:
        AddPackageRequest:
            properties:
                ecosystem:
                    type: string
                identifier:
                    type: string
            type: object
        AddPackageResponse:
            properties:
                already_exists:
                    type: boolean
                ecosystem:
                    type: string
                id:
                    format: int32
                    type: integer
                identifier:
                    type: string
            type: object
        CreatePackageRequest:
            properties:
                ecosystem:
                    type: string
                identifier:
                    type: string
            type: object
        CreatePackageResponse:
            properties:
                already_exists:
                    type: boolean
                ecosystem:
                    type: string
                id:
                    format: int32
                    type: integer
                identifier:
                    type: string
            type: object
...
        Ecosystem:
            enum:
                - npm
                - go
                - cargo
                - pypi
            type: string
        ErrorResponse:
            properties:
                error:
                    type: string
            required:
                - error
            type: object
...
        PackageListItem:
            properties:
                ecosystem:
                    type: string
                id:
                    format: int32
                    type: integer
                identifier:
                    type: string
                latest_version:
                    nullable: true
                    type: string
            type: object
        PackageListResponse:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/PackageListItem'
                    type: array
            type: object
        PackageSummary:
            properties:
                ecosystem:
                    $ref: '#/components/schemas/Ecosystem'
                identifier:
                    type: string
                latest_version:
                    type: string
            type: object
        PackageVersion:
            properties:
                ecosystem:
                    $ref: '#/components/schemas/Ecosystem'
                identifier:
                    type: string
                latest:
                    type: boolean
                maintainer_notes:
                    type: string
                source:
                    $ref: '#/components/schemas/Source'
                tags:
                    items:
                        $ref: '#/components/schemas/TagInfo'
                    type: array
                trust_level:
                    format: int32
                    type: integer
                version:
                    type: string
            type: object
        PkgVtype:
            enum:
                - integer
                - boolean
                - float
            type: string
...
        SearchResult:
            properties:
                items:
                    items:
                        $ref: '#/components/schemas/PackageSummary'
                    type: array
            type: object
        Source:
            properties:
                commit:
                    type: string
                tag:
                    type: string
                url:
                    type: string
            type: object
...
        TagInfo:
            properties:
                data:
                    format: byte
                    type: string
                label:
                    type: string
                value_type:
                    $ref: '#/components/schemas/PkgVtype'
            type: object
...
info:
    description: API for querying package information and metadata
    title: Secure Package Registry API
    version: 1.0.0
openapi: 3.0.3
paths:
    /api/v1/admin/packages:
        get:
            description: List all watched packages for a given ecosystem.
            operationId: adminListPackages
            parameters:
                - description: Ecosystem to list packages for
                  in: query
                  name: ecosystem
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PackageListResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: List watched packages
            tags:
                - Admin
        post:
            description: Create a new package in the watch list and publish a package requested event.
            operationId: adminAddPackage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AddPackageRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AddPackageResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Add package to watch list
            tags:
                - Admin
...
    /api/v1/internal/packages:
        post:
            description: Register a package from an internal service. Returns existing package if already created.
            operationId: internalCreatePackage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/CreatePackageRequest'
                required: true
            responses:
                "201":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/CreatePackageResponse'
                    description: Created
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Create package (internal)
            tags:
                - Internal
...
    /api/v1/svc/packages:
        get:
            description: Search for packages by query string. Returns all packages (paginated) if no query is provided.
            operationId: searchPackages
            parameters:
                - description: Search query string
                  in: query
                  name: q
                  schema:
                    type: string
                - description: Filter by ecosystem (npm, go, cargo, pypi)
                  in: query
                  name: ecosystem
                  schema:
                    type: string
                - description: Page number (default 1)
                  in: query
                  name: page
                  schema:
                    type: string
                - description: Page size (default 20, max 100)
                  in: query
                  name: page_size
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/SearchResult'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Search packages
            tags:
                - Packages
    /api/v1/svc/packages/{ecosystem}/{identifier}/{version}:
        get:
            description: Retrieve details for a specific package version.
            operationId: getPackageVersion
            parameters:
                - in: path
                  name: ecosystem
                  required: true
                  schema:
                    type: string
                - in: path
                  name: identifier
                  required: true
                  schema:
                    type: string
                - in: path
                  name: version
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/PackageVersion'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Bad Request
                "404":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Not Found
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ErrorResponse'
                    description: Internal Server Error
            summary: Get package version
            tags:
                - Packages
...